refactor: refactor store.ts
