+++ next/src/store/index.ts
@@ -1,0 +1,6 @@
+export { useCategoryStore } from './categoryStore'
+export { useSouvenirStore } from './souvenirStore'
+export { useCurrencyStore } from './currencyStore'
+export { useFavoriteStore } from './favoriteStore'
+export { useMyPostsStore } from './postStore'
+export { useCurrentUserStore } from './userStore'
